# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  userId: string;
  fullName: string;
  bloodGroup: string;
  allergies?: string;
  medications?: string;
  emergencyContacts?: string;
  insuranceInfo?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface UserProfileInput {
  userId: string;
  fullName: string;
  bloodGroup: string;
  allergies?: string;
  medications?: string;
  emergencyContacts?: string;
  insuranceInfo?: string;
}

export interface SosAlert {
  id: number;
  userId: string;
  latitude?: number;
  longitude?: number;
  status: string;
  triggeredAt: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
  severity?: string;
  /** @nullable */
  notes?: string | null;
}

export interface SosInput {
  userId: string;
  latitude?: number;
  longitude?: number;
  severity?: string;
  notes?: string;
}

export interface ActiveSosStatus {
  active: boolean;
  alert?: SosAlert;
}

export interface Incident {
  id: number;
  userId: string;
  type: string;
  description?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  severity: string;
  status: string;
  createdAt: string;
}

export interface IncidentInput {
  userId: string;
  type: string;
  description?: string;
  latitude?: number;
  longitude?: number;
  severity: string;
}

export interface ChatMessage {
  id: number;
  userId: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatInput {
  userId: string;
  message: string;
}

export interface EmergencyStats {
  severityScore: number;
  ambulanceEta: number;
  nearbyTraumaCenters: number;
  activeResponders: number;
  networkQuality: string;
  emergencyReadiness: number;
  gpsStatus: string;
  totalSosThisMonth?: number;
  avgResponseTime?: number;
}

export interface NearbyService {
  id: string;
  name: string;
  type: string;
  latitude?: number;
  longitude?: number;
  distance: number;
  phone?: string;
  eta?: number;
  available: boolean;
}

export type GetNearbyServicesParams = {
lat?: number;
lng?: number;
type?: string;
};