# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessage {
  id: number;
  userId: string;
  role: string;
  content: string;
  createdAt: string;
}