# lib/api-zod/src/generated/types/emergencyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyStats {
  severityScore: number;
  ambulanceEta: number;
  nearbyTraumaCenters: number;
  activeResponders: number;
  networkQuality: string;
  emergencyReadiness: number;
  gpsStatus: string;
  totalSosThisMonth?: number;
  avgResponseTime?: number;
}