# lib/api-zod/src/generated/types/incident.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface Incident {
  id: number;
  userId: string;
  type: string;
  description?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  severity: string;
  status: string;
  createdAt: string;
}