# lib/api-zod/src/generated/types/nearbyService.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface NearbyService {
  id: string;
  name: string;
  type: string;
  latitude?: number;
  longitude?: number;
  distance: number;
  phone?: string;
  eta?: number;
  available: boolean;
}