# lib/api-zod/src/generated/types/sosAlert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface SosAlert {
  id: number;
  userId: string;
  latitude?: number;
  longitude?: number;
  status: string;
  triggeredAt: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  resolvedAt?: string | null;
  severity?: string;
  /** @nullable */
  notes?: string | null;
}