# lib/api-zod/src/generated/types/sosInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface SosInput {
  userId: string;
  latitude?: number;
  longitude?: number;
  severity?: string;
  notes?: string;
}