# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  userId: string;
  fullName: string;
  bloodGroup: string;
  allergies?: string;
  medications?: string;
  emergencyContacts?: string;
  insuranceInfo?: string;
  createdAt?: string;
  updatedAt?: string;
}