# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ROADSOS AI Emergency Response API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  userId: string;
  fullName: string;
  bloodGroup: string;
  allergies?: string;
  medications?: string;
  emergencyContacts?: string;
  insuranceInfo?: string;
}